Move config to app/point
