+++ app/point/config/config.go
