+++ app/router/router.go
@@ -1,8 +1,8 @@
 package router
 
 import (
+	"github.com/v2ray/v2ray-core/app/point/config"
 	v2net "github.com/v2ray/v2ray-core/common/net"
-	"github.com/v2ray/v2ray-core/config"
 )
 
 type Router interface {
